Small Commit

+++ Borrower-Information-Microservice/src/main/java/com/lms/BI/Repository/BorrowerDetailsRepository.java
@@ -4,9 +4,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.lms.BI.Model.BorrowerDetails;
 
-public interface BorrowerDetailsRepository extends JpaRepository<BorrowerDetails, Long> {
+public interface BorrowerDetailsRepository extends JpaRepository<BorrowerDetails, Integer> {
     
-    BorrowerDetails findById(long id);  
+    BorrowerDetails findById(int id);  
     
     BorrowerDetails findByUname(String uname);
